Extract ISO_8859_2 to its own file
