+++ escpos4k/src/commonMain/kotlin/cz/multiplatform/escpos4k/core/encoding/charset/ISO_8859_2.kt
@@ -1,0 +1,152 @@
+/*
+ *    Copyright 2024 Ondřej Karmazín
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package cz.multiplatform.escpos4k.core.encoding.charset
+
+/** Page 39 [ISO8859-2: Latin 2] */
+public data object ISO_8859_2 : Charset(39, "ISO-8859-2") {
+  override val mapping: Map<Char, Int> =
+      mapOf(
+          '\u0080' to 0x80,
+          '\u0081' to 0x81,
+          '\u0082' to 0x82,
+          '\u0083' to 0x83,
+          '\u0084' to 0x84,
+          '\u0085' to 0x85,
+          '\u0086' to 0x86,
+          '\u0087' to 0x87,
+          '\u0088' to 0x88,
+          '\u0089' to 0x89,
+          '\u008A' to 0x8A,
+          '\u008B' to 0x8B,
+          '\u008C' to 0x8C,
+          '\u008D' to 0x8D,
+          '\u008E' to 0x8E,
+          '\u008F' to 0x8F,
+          '\u0090' to 0x90,
+          '\u0091' to 0x91,
+          '\u0092' to 0x92,
+          '\u0093' to 0x93,
+          '\u0094' to 0x94,
+          '\u0095' to 0x95,
+          '\u0096' to 0x96,
+          '\u0097' to 0x97,
+          '\u0098' to 0x98,
+          '\u0099' to 0x99,
+          '\u009A' to 0x9A,
+          '\u009B' to 0x9B,
+          '\u009C' to 0x9C,
+          '\u009D' to 0x9D,
+          '\u009E' to 0x9E,
+          '\u009F' to 0x9F,
+          '\u00A0' to 0xA0,
+          '\u0104' to 0xA1,
+          '\u02D8' to 0xA2,
+          '\u0141' to 0xA3,
+          '\u00A4' to 0xA4,
+          '\u013D' to 0xA5,
+          '\u015A' to 0xA6,
+          '\u00A7' to 0xA7,
+          '\u00A8' to 0xA8,
+          '\u0160' to 0xA9,
+          '\u015E' to 0xAA,
+          '\u0164' to 0xAB,
+          '\u0179' to 0xAC,
+          '\u00AD' to 0xAD,
+          '\u017D' to 0xAE,
+          '\u017B' to 0xAF,
+          '\u00B0' to 0xB0,
+          '\u0105' to 0xB1,
+          '\u02DB' to 0xB2,
+          '\u0142' to 0xB3,
+          '\u00B4' to 0xB4,
+          '\u013E' to 0xB5,
+          '\u015B' to 0xB6,
+          '\u02C7' to 0xB7,
+          '\u00B8' to 0xB8,
+          '\u0161' to 0xB9,
+          '\u015F' to 0xBA,
+          '\u0165' to 0xBB,
+          '\u017A' to 0xBC,
+          '\u02DD' to 0xBD,
+          '\u017E' to 0xBE,
+          '\u017C' to 0xBF,
+          '\u0154' to 0xC0,
+          '\u00C1' to 0xC1,
+          '\u00C2' to 0xC2,
+          '\u0102' to 0xC3,
+          '\u00C4' to 0xC4,
+          '\u0139' to 0xC5,
+          '\u0106' to 0xC6,
+          '\u00C7' to 0xC7,
+          '\u010C' to 0xC8,
+          '\u00C9' to 0xC9,
+          '\u0118' to 0xCA,
+          '\u00CB' to 0xCB,
+          '\u011A' to 0xCC,
+          '\u00CD' to 0xCD,
+          '\u00CE' to 0xCE,
+          '\u010E' to 0xCF,
+          '\u0110' to 0xD0,
+          '\u0143' to 0xD1,
+          '\u0147' to 0xD2,
+          '\u00D3' to 0xD3,
+          '\u00D4' to 0xD4,
+          '\u0150' to 0xD5,
+          '\u00D6' to 0xD6,
+          '\u00D7' to 0xD7,
+          '\u0158' to 0xD8,
+          '\u016E' to 0xD9,
+          '\u00DA' to 0xDA,
+          '\u0170' to 0xDB,
+          '\u00DC' to 0xDC,
+          '\u00DD' to 0xDD,
+          '\u0162' to 0xDE,
+          '\u00DF' to 0xDF,
+          '\u0155' to 0xE0,
+          '\u00E1' to 0xE1,
+          '\u00E2' to 0xE2,
+          '\u0103' to 0xE3,
+          '\u00E4' to 0xE4,
+          '\u013A' to 0xE5,
+          '\u0107' to 0xE6,
+          '\u00E7' to 0xE7,
+          '\u010D' to 0xE8,
+          '\u00E9' to 0xE9,
+          '\u0119' to 0xEA,
+          '\u00EB' to 0xEB,
+          '\u011B' to 0xEC,
+          '\u00ED' to 0xED,
+          '\u00EE' to 0xEE,
+          '\u010F' to 0xEF,
+          '\u0111' to 0xF0,
+          '\u0144' to 0xF1,
+          '\u0148' to 0xF2,
+          '\u00F3' to 0xF3,
+          '\u00F4' to 0xF4,
+          '\u0151' to 0xF5,
+          '\u00F6' to 0xF6,
+          '\u00F7' to 0xF7,
+          '\u0159' to 0xF8,
+          '\u016F' to 0xF9,
+          '\u00FA' to 0xFA,
+          '\u0171' to 0xFB,
+          '\u00FC' to 0xFC,
+          '\u00FD' to 0xFD,
+          '\u0163' to 0xFE,
+          '\u02D9' to 0xFF,
+      )
+}
